requirements.txt
bs4
psycopg2-binary
requests
apscheduler
schedule